requirements.txt
boto3 == 1.37.38
pandas == 2.0.3
psycopg2-binary == 2.9.10
python-dotenv == 1.0.1

